Mastermind refactor - add game.rb
