+++ test/player_guess_matcher_test.rb
@@ -6,35 +6,44 @@ require './lib/player_guess_matcher'
 require 'pry'
 
 class PlayerGuessMatcherTest < Minitest::Test
-  # def test_it_exists
-  #   # guess_count = 1
-  #   # winning_code = 'gryb'
-  #   # player_guess = 'rryg'
-  #   # mastermind = Mastermind.new(guess_count, winning_code)
-  #   player_guess_matcher = PlayerGuessMatcher.new(winning_code, player_guess)
-  #   binding.pry
-  #   assert_instance_of PlayerGuessMatcher, player_guess_matcher
-  # end
+  def test_it_exists
+    guess_count = 1
+    winning_code = 'gryb'
+
+    player_guess_matcher = PlayerGuessMatcher.new(winning_code, guess_count)
+
+    assert_instance_of PlayerGuessMatcher, player_guess_matcher
+  end
 
   def test_readable_attributes
-    # mastermind = Mastermind.new
-    player_guess_matcher = PlayerGuessMatcher.new(@winning_code, @player_guess)
+
+    player_guess_matcher = PlayerGuessMatcher.new(@winning_code, @guess_count)
 
     assert_equal @winning_code, player_guess_matcher.winning_code
   end
 
   def test_check_number_of_correct_colors
 
-    winning_code = 'ygry'
-    player_guess = 'bryg'
+    @winning_code = 'ygry'
+    @player_guess = 'bryg'
 
-    mastermind = Mastermind.new(@guess_count, winning_code)
-    player_guess_matcher = PlayerGuessMatcher.new(winning_code, player_guess)
+    mastermind = Mastermind.new(@winning_code)
+    player_guess_matcher = PlayerGuessMatcher.new(@winning_code, @player_guess)
 
-    assert_equal 3, player_guess_matcher.check_player_guess(player_guess)
+    assert_equal 3, player_guess_matcher.check_number_of_correct_colors(@player_guess)
   end
 
   def test_check_number_of_correct_color_and_position
 
+    @winning_code = 'bgry'
+    @player_guess = 'brrg'
+
+    @winning_code_array = ['b', 'g', 'r', 'y']
+    @player_guess_array = ['b', 'r', 'r', 'g']
+
+    mastermind = Mastermind.new(@winning_code)
+    player_guess_matcher = PlayerGuessMatcher.new(@winning_code, @player_guess)
+
+    assert_equal 1, player_guess_matcher.check_number_of_correct_color_and_position(@player_guess)
   end
 end
